+++ storehaus-hbase/src/main/scala/com/twitter/storehaus/hbase/HBaseLongStore.scala
@@ -18,7 +18,7 @@ package com.twitter.storehaus.hbase
 
 import org.apache.hadoop.hbase.client.HTablePool
 import com.twitter.storehaus.Store
-import com.twitter.util.Future
+import com.twitter.util.{Future, Time}
 import com.twitter.bijection.Injection._
 import org.apache.hadoop.conf.Configuration
 
@@ -81,7 +81,7 @@ class HBaseLongStore(protected val quorumNames: Seq[String],
   /** Close this store and release any resources.
     * It is undefined what happens on get/multiGet after close
     */
-  override def close {
+  override def close(t: Time) = futurePool {
     pool.close()
   }
 }